# public/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Znaidi Mahdi Starter - Open API",
        "description": "APIs documentation usin OpenAPI",
        "contact": {
            "name": "Znaidi Mahdi",
            "email": "mahdi.znaidi@esprit.tn"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "a local dev server"
        }
    ],
    "paths": {
...
        "/postes": {
            "get": {
                "tags": [
                    "Poste"
                ],
                "summary": "List all Postes",
                "operationId": "indexPoste",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "List All Postes",
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "code": {
                                            "description": "List All Postes",
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "description": "List All Postes",
                                            "type": "string",
                                            "example": "postes_list"
                                        },
                                        "data": {
                                            "description": "List All Postes",
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Poste"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "Password Based": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Poste"
                ],
                "summary": "Create a new Poste",
                "operationId": "storePoste",
                "requestBody": {
                    "$ref": "#/components/requestBodies/Poste"
                },
                "responses": {
                    "201": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Create a new Poste",
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "code": {
                                            "description": "Create a new Poste",
                                            "type": "integer",
                                            "example": 201
                                        },
                                        "message": {
                                            "description": "Create a new Poste",
                                            "type": "string",
                                            "example": "poste_created"
                                        },
                                        "data": {
                                            "description": "Create a new Poste",
                                            "properties": {
                                                "id": {
                                                    "description": "Create a new Poste",
                                                    "type": "integer"
                                                },
                                                "libelle_poste": {
                                                    "description": "Create a new Poste",
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "Password Based": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Poste": {
                "title": "Poste",
                "description": "Poste model",
                "required": [
                    "nom"
                ],
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "Poste ID",
                        "type": "integer",
                        "format": "int"
                    },
                    "libelle_poste": {
                        "title": "libelle_poste",
                        "description": "Poste libelle",
                        "type": "string",
                        "format": "string"
                    },
                    "users": {
                        "title": "users",
                        "description": "Poste users",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/User"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Poste"
                }
            },
...
            "User": {
                "title": "User",
                "description": "User model",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "User id",
                        "type": "integer",
                        "format": "integer"
                    },
                    "email": {
                        "title": "email",
                        "description": "User email",
                        "type": "string",
                        "format": "string"
                    },
                    "matricule_hr": {
                        "title": "matricule_hr",
                        "description": "User matricule_hr",
                        "type": "string",
                        "format": "string"
                    },
                    "nom": {
                        "title": "nom",
                        "description": "User nom",
                        "type": "string",
                        "format": "string"
                    },
                    "prenom": {
                        "title": "prenom",
                        "description": "User prenom",
                        "type": "string",
                        "format": "string"
                    },
                    "adresse": {
                        "title": "adresse",
                        "description": "User adresse",
                        "type": "string",
                        "format": "string"
                    },
                    "num_telephone": {
                        "title": "numéro_téléphone",
                        "description": "User numéro_telephone",
                        "type": "string",
                        "format": "string"
                    },
                    "cin_passport": {
                        "title": "cin_passport",
                        "description": "User cin_passport",
                        "type": "string",
                        "format": "string"
                    },
                    "etat_civil": {
                        "title": "etat_civil",
                        "description": "User etat_civil",
                        "type": "string",
                        "format": "string"
                    },
                    "image": {
                        "title": "image",
                        "description": "User image",
                        "type": "string",
                        "format": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User"
                }
            },
...
            "PosteRequestBody": {
                "title": "PostRequestBody",
                "description": "Poste model",
                "properties": {
                    "libelle_poste": {
                        "title": "libelle_poste",
                        "description": "Poste libelle",
                        "type": "string",
                        "format": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Post"
                }
            },
...
        },
        "requestBodies": {
...
            "Poste": {
                "description": "Poste object that needs to be added to the store",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/PosteRequestBody"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "Password Based": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}